privatize the section root block client id selector
Conflicts:
	packages/block-editor/src/store/private-selectors.js
